requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
Flask-JWT-Extended>=4.6,<5.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<24.0
pytest>=8.0,<9.0
